requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
python-dotenv==1.0.1
requests==2.31.0
flask-swagger-ui==4.11.1
elastic-transport==8.13.1
elasticsearch==8.14.0
elasticsearch-async==6.2.0
llama_models==0.0.42
llama_stack==0.0.42
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
huggingface-hub==0.25.2